requirements.txt
numpy==1.22.2
replit==3.2.4
Flask==2.2.0
urllib3==1.26.6
PyNaCl==1.5.0
discord.py==2.2.2
pafy==0.5.5
ffmpeg==1.4
opuslib==3.0.1
python-dotenv==1.0.0
pytz==2023.3.*
audiofile
pymongo